Drop notification bell; promote notifications to /account tab + always-on toast watcher.
User feedback: the bell felt out of place in the navbar. Toasts now surface live arrivals; the durable list lives on /account?tab=notifications. Notification helpers (describe/summarize/href) moved to components/notifications/ so the toast watcher and the tab share one source of truth. The watcher also re-polls on visibilitychange so a user returning to the tab sees anything queued in the interim without waiting out the 30s interval.
